z80-asm: Fix parsing of chars and quotes in strings.
In defb expressions, things like "abc"+5 were supported,
which added 5 to the last character. I don't know if that's
useful, and the parser got wrong answers for "D"-32+64
(actually returned "D"-(32+64)), so I removed it and
now you can only do math on single characters.

Also doubling the quote character works as that quote char
in both strings and chars: 'that''s all folks!'