trs80-emulator-web: Try SVG screen.
This is an experiment to see if I can use an SVG image
instead of a canvas for the TRS-80 screen. I mostly have it
looking pretty good, but I occasionally see artifacts with
the blur, and overall I feel insecure about using this method
because I'm at the mercy of browsers' rendering bugs.

The ideas was to use SVG so I could try the various filters
and get a CRT effect. If I'm serious about that I should just
use WebGL and have full control over every pixel.

Another downside of SVG is that the user can't just download
the image as a PNG. There may be other downsides, like we
can't take screenshots for program icons, or copy the
selection to the clipboard (in the IDE).